Friday, September 14 2012
The EBU media player and his features have been improved. Player is more
stable. Exceptions could yet happen because there is not really control
on what kind of files we are trying to open.
A playlist was added to the player with M3U support. Playlist basic
functions are implemented.
Metadata Viewer is not enough stable yet. If we try to open an MXF
without EBUCore Metadata, it crashes, if you try to expand metadata tree
to quickly, it crashes, etc...